+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Version 9.0.0-dev
 - BC break: removed prefix I from interfaces' names
 - argument --filterExceptFolders in mytester.php now applies also to .phpt files
 - BC break: replaced mytester.php's arguments --coverageFormat and --coverageFile with --coverage, --resultsFormat and --resultsFile with --results
-- made it possible to generate multiple code coverage reports at once in mytester.php
+- made it possible to generate multiple (code coverage) reports at once in mytester.php
 - BC break: Job::$callback accepts only Closure (first class callable syntax) now
 - BC break: removed TestCase::METHOD_PATTERN